refactor(honeycomb): move HoneycombIntegrationConfig into vendor-first config module
Relocate HoneycombIntegrationConfig and its DEFAULT_HONEYCOMB_* constants
out of the shared integrations/config_models.py into
integrations/honeycomb/config.py, so the Honeycomb package is fully
self-contained ({config, client, verifier, tools}). All import sites are
rewritten to the canonical vendor path with no forwarding shim, matching
the existing airflow.config precedent in _catalog_impl.py.

First exemplar for T-18 (#3551) vendor-first completion.

Refs #3551
