+++ integrations/config_models.py
@@ -23,8 +23,6 @@ _LOCAL_GRAFANA_HOSTS = {"localhost", "127.0.0.1", "0.0.0.0"}
 DEFAULT_GROUNDCOVER_MCP_URL = "https://mcp.groundcover.com/api/mcp"
 DEFAULT_GROUNDCOVER_TIMEZONE = "UTC"
 DEFAULT_DATADOG_SITE = "datadoghq.com"
-DEFAULT_HONEYCOMB_BASE_URL = "https://api.honeycomb.io"
-DEFAULT_HONEYCOMB_DATASET = "__all__"
 DEFAULT_CORALOGIX_BASE_URL = "https://api.coralogix.com"
 DEFAULT_OPSGENIE_BASE_URLS: dict[str, str] = {
     "us": "https://api.opsgenie.com",
@@ -196,22 +194,6 @@ class GroundcoverIntegrationConfig(StrictConfigModel):
         if self.backend_id:
             headers["X-Backend-Id"] = self.backend_id
         return headers
-
-
-class HoneycombIntegrationConfig(StrictConfigModel):
-    """Normalized Honeycomb credentials used by resolution and verification flows."""
-
-    api_key: str
-    dataset: str = DEFAULT_HONEYCOMB_DATASET
-    base_url: str = DEFAULT_HONEYCOMB_BASE_URL
-    integration_id: str = ""
-
-    _normalize_dataset = field_validator("dataset", mode="before")(
-        normalize_with_default(DEFAULT_HONEYCOMB_DATASET)
-    )
-    _normalize_base_url = field_validator("base_url", mode="before")(
-        normalize_url(DEFAULT_HONEYCOMB_BASE_URL)
-    )
 
 
 class CoralogixIntegrationConfig(StrictConfigModel):
