+++ integrations/honeycomb/verifier.py
@@ -2,8 +2,8 @@
 
 from __future__ import annotations
 
-from integrations.config_models import HoneycombIntegrationConfig
 from integrations.honeycomb.client import HoneycombClient
+from integrations.honeycomb.config import HoneycombIntegrationConfig
 from integrations.verification import register_probe_verifier
 
 verify_honeycomb = register_probe_verifier(
